[#5] refactor: ContentView 내 전역 의존성 주입
